backend/requirements.txt
bcrypt==4.1.2
blinker==1.7.0
cachelib==0.12.0
click==8.1.7
colorama==0.4.6
Flask==2.3.3
Flask-Login==0.6.3
Flask-Session==0.6.0
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
MarkupSafe==2.1.5
mysql-connector-python==8.3.0
mysqlclient==2.2.4
numpy
pandas
python-dateutil==2.9.0.post0
pytz==2024.1
scikit-learn
scipy==1.12.0
six==1.16.0
threadpoolctl==3.3.0
tzdata==2024.1
Werkzeug==3.0.1